requirements.txt
Django==1.5
dj-database-url==0.3.0
gunicorn==19.1.1
psycopg2==2.5.1
python-social-auth
python-twitter==2.2
South==1.0.2
Fabric==1.7.0
hachoir-core
hachoir-parser
hachoir-metadata
